pyproject.toml
[tool.poetry]
name = "django-2-llunareine"
version = "0.1.0"
description = ""
authors = ["Aiganym <llunareine@gmail.com>"]
readme = "README.md"
packages = [{include = "django_2_llunareine"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
django-debug-toolbar = "^4.2.0"
djangorestframework = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
